refactor `Assembler::resolve_target()`
