+++ assembly/src/assembler/instruction/procedures.rs
@@ -1,11 +1,11 @@
 use smallvec::SmallVec;
-use vm_core::{mast::MastNodeId, utils::Either};
+use vm_core::mast::MastNodeId;
 
 use super::{Assembler, BasicBlockBuilder, Operation};
 use crate::{
     assembler::{mast_forest_builder::MastForestBuilder, ProcedureContext},
     ast::{InvocationTarget, InvokeKind},
-    AssemblyError, RpoDigest, SourceSpan, Spanned,
+    AssemblyError, RpoDigest,
 };
 
 /// Procedure Invocation
@@ -17,88 +17,14 @@ impl Assembler {
         proc_ctx: &mut ProcedureContext,
         mast_forest_builder: &mut MastForestBuilder,
     ) -> Result<MastNodeId, AssemblyError> {
-        let span = callee.span();
-        let node_id_or_digest = self.resolve_target(kind, callee, proc_ctx, mast_forest_builder)?;
-        let invoked_proc_node_id = match node_id_or_digest {
-            Either::Left(node_id) => node_id,
-            Either::Right(mast_root) => {
-                self.get_proc_root_id_from_mast_root(kind, span, mast_root, mast_forest_builder)?
-            },
-        };
+        let invoked_proc_node_id =
+            self.resolve_target(kind, callee, proc_ctx, mast_forest_builder)?;
 
         match kind {
             InvokeKind::ProcRef | InvokeKind::Exec => Ok(invoked_proc_node_id),
             InvokeKind::Call => mast_forest_builder.ensure_call(invoked_proc_node_id),
             InvokeKind::SysCall => mast_forest_builder.ensure_syscall(invoked_proc_node_id),
         }
-    }
-
-    /// Returns the [`MastNodeId`] associated with the provided MAST root if known, or wraps the
-    /// MAST root in a [`vm_core::mast::ExternalNode`] and returns it.
-    fn get_proc_root_id_from_mast_root(
-        &self,
-        kind: InvokeKind,
-        span: SourceSpan,
-        mast_root: RpoDigest,
-        mast_forest_builder: &mut MastForestBuilder,
-    ) -> Result<MastNodeId, AssemblyError> {
-        // Get the procedure from the assembler
-        let current_source_file = self.source_manager.get(span.source_id()).ok();
-
-        // If the procedure is cached and is a system call, ensure that the call is valid.
-        match mast_forest_builder.find_procedure_by_mast_root(&mast_root) {
-            Some(proc) if matches!(kind, InvokeKind::SysCall) => {
-                // Verify if this is a syscall, that the callee is a kernel procedure
-                //
-                // NOTE: The assembler is expected to know the full set of all kernel
-                // procedures at this point, so if we can't identify the callee as a
-                // kernel procedure, it is a definite error.
-                if !proc.visibility().is_syscall() {
-                    return Err(AssemblyError::InvalidSysCallTarget {
-                        span,
-                        source_file: current_source_file,
-                        callee: proc.fully_qualified_name().clone(),
-                    });
-                }
-                let maybe_kernel_path = proc.path();
-                self.module_graph
-                    .find_module(maybe_kernel_path)
-                    .ok_or_else(|| AssemblyError::InvalidSysCallTarget {
-                        span,
-                        source_file: current_source_file.clone(),
-                        callee: proc.fully_qualified_name().clone(),
-                    })
-                    .and_then(|module| {
-                        // Note: this module is guaranteed to be of AST variant, since we have the
-                        // AST of a procedure contained in it (i.e. `proc`). Hence, it must be that
-                        // the entire module is in AST representation as well.
-                        if module.unwrap_ast().is_kernel() {
-                            Ok(())
-                        } else {
-                            Err(AssemblyError::InvalidSysCallTarget {
-                                span,
-                                source_file: current_source_file.clone(),
-                                callee: proc.fully_qualified_name().clone(),
-                            })
-                        }
-                    })?;
-            },
-            Some(_) | None => (),
-        }
-
-        // Note that here we rely on the fact that we topologically sorted the procedures, such that
-        // when we assemble a procedure, all procedures that it calls will have been assembled, and
-        // hence be present in the `MastForest`.
-        let invoked_node_id = match mast_forest_builder.find_procedure_node_id(mast_root) {
-            Some(root) => root,
-            None => {
-                // If the MAST root called isn't known to us, make it an external
-                // reference.
-                mast_forest_builder.ensure_external(mast_root)?
-            },
-        };
-
-        Ok(invoked_node_id)
     }
 
     /// Creates a new DYN block for the dynamic code execution and return.
@@ -129,18 +55,15 @@ impl Assembler {
         callee: &InvocationTarget,
         proc_ctx: &mut ProcedureContext,
         basic_block_builder: &mut BasicBlockBuilder,
-        mast_forest_builder: &MastForestBuilder,
+        mast_forest_builder: &mut MastForestBuilder,
     ) -> Result<(), AssemblyError> {
-        let digest = match self.resolve_target(
-            InvokeKind::ProcRef,
-            callee,
-            proc_ctx,
-            mast_forest_builder,
-        )? {
-            Either::Left(node_id) => mast_forest_builder.get_mast_node(node_id).unwrap().digest(),
-            Either::Right(digest) => digest,
+        let mast_root = {
+            let proc_body_id =
+                self.resolve_target(InvokeKind::ProcRef, callee, proc_ctx, mast_forest_builder)?;
+            mast_forest_builder.get_mast_node(proc_body_id).unwrap().digest()
         };
-        self.procref_mast_root(digest, basic_block_builder)
+
+        self.procref_mast_root(mast_root, basic_block_builder)
     }
 
     fn procref_mast_root(
